backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
chromadb==0.4.18
llama-cpp-python==0.2.23
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3 